CLN: Refactor ObjectDataProvider
This splits the large elif cases into a factory pattern and establishes
a providers subpackage. Some extra refactoring is still warranted for
the case of existing metadata and some other things can be cleaned up
with this split.
